+++ crates/dioxus-class-internal/src/lib.rs
@@ -1,3 +1,5 @@
+#![doc = include_str!("../README.md")]
+
 pub mod class;
 
 pub use class::Class;